store: add podmanEndpoints to composerApi
