+++ src/store/api/backend/onprem/composerApi/index.ts
@@ -29,6 +29,7 @@ import { architectureEndpoints } from './architecture';
 import { blueprintEndpoints } from './blueprints';
 import { getBlueprintsPath } from './helpers';
 import { oscapEndpoints } from './oscap';
+import { podmanEndpoints } from './podman';
 import { workerEndpoints } from './worker';
 
 import { emptyComposerApi } from '../emptyComposerApi';
@@ -39,9 +40,6 @@ import {
   type ComposerImageRequest,
   type ComposerCustomizations as Customizations,
   type ComposerImageTypes as ImageTypes,
-  type PodmanImageInfo,
-  type PodmanImagesArg,
-  type PodmanImagesResponse,
 } from '../types';
 
 const readComposes = async (bpID: string) => {
@@ -156,6 +154,7 @@ export const composerApi = emptyComposerApi.injectEndpoints({
       ...blueprintEndpoints(builder),
       ...oscapEndpoints(builder),
       ...workerEndpoints(builder),
+      ...podmanEndpoints(builder),
       composeBlueprint: builder.mutation<
         ComposeBlueprintApiResponse,
         ComposeBlueprintApiArg
@@ -292,56 +291,6 @@ export const composerApi = emptyComposerApi.injectEndpoints({
           }
 
           throw new Error('Compose not found');
-        }),
-      }),
-      podmanImages: builder.query<PodmanImagesResponse, PodmanImagesArg>({
-        queryFn: onPremQueryHandler(async () => {
-          const result = (await cockpit.spawn(
-            [
-              'podman',
-              'images',
-              '--filter',
-              'reference=registry.redhat.io/rhel*/rhel-bootc',
-              '--format',
-              '{{.Repository}},{{.Tag}}',
-            ],
-            {
-              superuser: 'require',
-            },
-          )) as string;
-
-          if (!result.trim()) {
-            return [];
-          }
-
-          const images = result
-            .trim()
-            .split('\n')
-            .map((s) => {
-              if (!s.trim()) {
-                return null;
-              }
-
-              const parts = s.trim().split(',');
-              if (parts.length !== 2) {
-                // Skip malformed lines that don't match "repository,tag"
-                return null;
-              }
-
-              const [repository, tag] = parts.map((p) => p.trim());
-              if (!repository || !tag) {
-                return null;
-              }
-
-              return {
-                image: `${repository}:${tag}`,
-                repository,
-                tag,
-              };
-            })
-            .filter((image): image is PodmanImageInfo => image !== null);
-
-          return images;
         }),
       }),
     };
